requirements.txt
pandas==2.2.2
psycopg2==2.9.9
openpyxl==3.1.5
python-dotenv==1.0.1
pyyaml==6.0.2